pyproject.toml
[tool.poetry]
name = "llmeasy"
version = "0.1.0"
description = "A flexible framework for integrating multiple AI chat providers"
authors = ["Kevin Wong <kevinchwong@gmail.com>"]
license = "Apache License 2.0"
readme = "README.md"
packages = [{include = "llmeasy"}]

[tool.poetry.dependencies]
python = "^3.8"
openai = "^1.12.0"
anthropic = "^0.18.1"
google-generativeai = "^0.3.2"
mistralai = "^0.0.12"
python-dotenv = "^1.0.1"
aiohttp = "^3.9.3"
typing-extensions = "^4.9.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-asyncio = "^0.23.5"
black = "^24.1.1"
isort = "^5.13.2"
mypy = "^1.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"